Wishlist, WishItem, and WishlistController changes
WishItem now has an enumeration ItemStatus attribute to indicate the
order status of that specific item.
Wishlist now has a Status attribute to be changed based on the least
complete WishItem.ItemStatus.
WishlistController has some leftover code that might be useful, but
needs to be reworked in line with current project